Split read vs write actors
