+++ kanidmd/src/lib/actors/mod.rs
@@ -1,1 +1,2 @@
-pub mod v1;
+pub mod v1_read;
+pub mod v1_write;
