+++ kanidmd/src/lib/plugins/attrunique.rs
@@ -225,7 +225,9 @@ mod tests {
         let preload = vec![e];
 
         run_create_test!(
-            Err(OperationError::Plugin(PluginError::AttrUnique("duplicate value detected".to_string()))),
+            Err(OperationError::Plugin(PluginError::AttrUnique(
+                "duplicate value detected".to_string()
+            ))),
             preload,
             create,
             None,
@@ -253,7 +255,9 @@ mod tests {
         let preload = Vec::new();
 
         run_create_test!(
-            Err(OperationError::Plugin(PluginError::AttrUnique("ava already exists".to_string()))),
+            Err(OperationError::Plugin(PluginError::AttrUnique(
+                "ava already exists".to_string()
+            ))),
             preload,
             create,
             None,
@@ -294,7 +298,9 @@ mod tests {
         let preload = vec![ea, eb];
 
         run_modify_test!(
-            Err(OperationError::Plugin(PluginError::AttrUnique("duplicate value detected".to_string()))),
+            Err(OperationError::Plugin(PluginError::AttrUnique(
+                "duplicate value detected".to_string()
+            ))),
             preload,
             filter!(f_or!([f_eq(
                 "name",
@@ -339,7 +345,9 @@ mod tests {
         let preload = vec![ea, eb];
 
         run_modify_test!(
-            Err(OperationError::Plugin(PluginError::AttrUnique("ava already exists".to_string()))),
+            Err(OperationError::Plugin(PluginError::AttrUnique(
+                "ava already exists".to_string()
+            ))),
             preload,
             filter!(f_or!([
                 f_eq("name", PartialValue::new_iutf8s("testgroup_a")),
